app/requirements.txt
aiosmtplib==2.0.2
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.0.1
blinker==1.7.0
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
Faker==20.0.0
fastapi==0.104.1
fastapi-mail==1.4.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
pyasn1==0.5.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
